First debuggable revision
Had problems using Olimex PIC KIT3 to debug. It was complaining about
some settings that do not allow the debugger to run. It took me a long
time (1 week) to figure out what the problem was. I found the problem
by building a very simple project which the debugger could handle. The
problem turned out to be one of the config settings in system_init.c.